+++ R/resolved.R
@@ -19,7 +19,23 @@
 #' e.g. `while (!resolved(future)) Sys.sleep(5)`.
 #'
 #' @export
-resolved <- function(x, ...) UseMethod("resolved")
+resolved <- function(x, ...) {
+  ## Automatically update journal entries for Future object
+  if (inherits(future, "Future") &&
+      inherits(future$.journal, "FutureJournal")) {
+    start <- Sys.time()
+    on.exit({
+      appendToFutureJournal(x,
+        event = "resolved",
+        category = "querying",
+        start = start,
+        stop = Sys.time(),
+        skip = FALSE
+      )
+    })
+  }
+  UseMethod("resolved")
+}
 
 #' @export
 resolved.default <- function(x, ...) TRUE